Se refactorizan la respuesta de los controladores para mejor consumo de los endpoints

+++ src/main/java/com/tictac/demo/service/CursoProyectoService.java
@@ -19,11 +19,35 @@ public class CursoProyectoService {
     }
 
     public CursoProyecto createCursoProyecto(CursoProyecto cursoProyecto){
-        return cursoProyectoRepository.save(cursoProyecto);
+        if(cursoProyecto.getIdProyecto() == null || cursoProyecto.getIdProyecto().toString().trim().isEmpty() ||
+            cursoProyecto.getIdActividad() == null || cursoProyecto.getIdActividad().toString().trim().isEmpty()){
+            return null;
+        }else {
+            return cursoProyectoRepository.save(cursoProyecto);
+        }
     }
 
-    public void deleteCursoProyecto(Integer id){
-        cursoProyectoRepository.deleteById(id);
+    public String updateCursoProyecto(CursoProyecto cursoProyecto){
+        if(cursoProyectoRepository.existsById(cursoProyecto.getGrado())){
+            Optional<CursoProyecto> cp = cursoProyectoRepository.findById(cursoProyecto.getGrado());
+
+            cp.get().setIdProyecto(cursoProyecto.getIdProyecto());
+            cp.get().setIdActividad(cursoProyecto.getIdActividad());
+
+            cursoProyectoRepository.save(cp.get());
+            return "Proyecto del curso actualizado con éxito";
+        }else{
+            return null;
+        }
+    }
+
+    public String deleteCursoProyecto(Integer id){
+        if(cursoProyectoRepository.existsById(id)){
+            cursoProyectoRepository.deleteById(id);
+            return "Proyecto del curso eliminado con éxito";
+        }else{
+            return null;
+        }
     }
 
     public List<CursoProyecto> listCursoProyecto(){
